pyproject.toml
[project]
name = "Pandas-Practice"
version = "0.1.0"
description = "Practice with Pandas for the BJSS Interview"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.35.90",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pydantic>=2.10.4",
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.36",
]

[dependency-groups]
dev = [
    "moto>=5.0.25",
    "pre-commit>=4.0.1",
    "pytest>=8.3.4",
    "ruff>=0.8.4",
    "sqlfluff>=3.3.0",
]

[tool.pytest.ini_options]
pythonpath = "./tests"
addopts = ["--strict-markers", "--strict-config", "-v", "-ra", "-q"]
xfail_strict = true
testpaths = ["tests"]
log_cli = true
log_cli_level = "INFO"
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
minversion = "6.0"
filterwarnings = "ignore"
norecursedirs = ["docs", "build"]
# Specify a custom directory for pytest cache
cache_dir = ".pytest_cache"

[tool.ruff]
target-version = "py311"
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    # Specific Rules to ignore
    "C417", # https://docs.astral.sh/ruff/rules/unnecessary-map/
    "COM812", # https://docs.astral.sh/ruff/rules/missing-trailing-comma/
    "COM819", # https://docs.astral.sh/ruff/rules/prohibited-trailing-comma/
    "E203", # https://docs.astral.sh/ruff/rules/whitespace-before-punctuation/
    "E402", # https://docs.astral.sh/ruff/rules/module-import-not-at-top-of-file/
    "E902", # https://docs.astral.sh/ruff/rules/io-error/
    "FA102", # https://docs.astral.sh/ruff/rules/future-required-type-annotation
    "F541", # https://docs.astral.sh/ruff/rules/f-string-missing-placeholders/
    "ISC001", # https://docs.astral.sh/ruff/rules/single-line-implicit-string-concatenation/
    # ^^ Conflicts with COM812
    "N999", #  https://docs.astral.sh/ruff/rules/invalid-module-name/
    "PLR0913", # https://docs.astral.sh/ruff/rules/too-many-arguments/
    "PLW0120", # https://docs.astral.sh/ruff/rules/useless-else-on-loop/#useless-else-on-loop-plw0120
    "RUF013", # https://docs.astral.sh/ruff/rules/implicit-optional/
    "T203", # https://docs.astral.sh/ruff/rules/p-print/
    "TCH002", # https://docs.astral.sh/ruff/rules/typing-only-third-party-import/
    "TCH003", # https://docs.astral.sh/ruff/rules/typing-only-standard-library-import/
    "TRY003", # https://docs.astral.sh/ruff/rules/raise-vanilla-args/

    # Rule-sets to ignore
    "ANN", # https://docs.astral.sh/ruff/rules/#flake8-annotations-ann
    "D", # https://docs.astral.sh/ruff/rules/#pydocstyle-d
    "DOC", # https://docs.astral.sh/ruff/rules/#pydoclint-doc
    "EM", # https://docs.astral.sh/ruff/rules/#flake8-errmsg-em
    "ERA", # https://docs.astral.sh/ruff/rules/#eradicate-era
    "G", # https://docs.astral.sh/ruff/rules/#flake8-logging-format-g
    "TD", # https://docs.astral.sh/ruff/rules/#flake8-todos-td
    "FIX" # https://docs.astral.sh/ruff/rules/#flake8-fixme-fix
]

fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.sqlfluff.core]
templater = "jinja"
sql_file_exts = ".sql,.sql.j2,.dml,.ddl"
max_line_length = 120
